+++ ui-main-screen/src/main/java/com/example/ui_main_screen/MainScreenEffects.kt
@@ -2,5 +2,5 @@ package com.example.ui_main_screen
 
 sealed interface MainScreenEffects {
     object NavigateToNewNote: MainScreenEffects
-    data class NavigateToNote(val noteId: Int): MainScreenEffects
+    data class NavigateToNote(val noteId: Long): MainScreenEffects
 }
