+++ ui-main-screen/src/main/java/com/example/ui_main_screen/MainScreenEvents.kt
@@ -2,5 +2,5 @@ package com.example.ui_main_screen
 
 sealed interface MainScreenEvents {
     object OnNewNoteClicked: MainScreenEvents
-    data class OnNoteClicked(val noteId: Int): MainScreenEvents
+    data class OnNoteClicked(val noteId: Long): MainScreenEvents
 }
